Move VueJs module template under SPA
