--- generators/spa-vue/templates/Module.dnn
@@ -1,102 +1,0 @@
-<dotnetnuke type="Package" version="5.0">
-  <packages>
-    <package name="<%= fullNamespace %>" type="Module" version="00.00.01">
-      <friendlyName><%= extensionName %></friendlyName>
-      <description>
-        <![CDATA[<p><%= description %></p>]]>
-      </description>
-      <iconFile>~/Icons/Sigma/Software_32X32_Standard.png</iconFile>
-      <owner>
-        <name><%= yourName %></name>
-        <organization><%= company %></organization>
-        <url><![CDATA[<a href="<%= companyUrl %>" target="_blank"><%= companyUrl %></a>]]></url>
-        <email><![CDATA[<a href="mailto:<%= emailAddy %>" target="_blank"><%= emailAddy %></a>]]></email>
-      </owner>
-      <license src="License.txt"></license>
-      <releaseNotes src="ReleaseNotes.txt"></releaseNotes>
-      <dependencies>
-        <dependency type="CoreVersion">09.03.02</dependency>
-      </dependencies>
-      
-      <components>
-        <component type="Script">
-          <scripts>
-            <basePath>DesktopModules\<%= extensionName %></basePath>
-            <script type="Install">
-              <path>Providers\DataProviders\SqlDataProvider</path>
-              <name>00.00.01.SqlDataProvider</name>
-              <version>00.00.01</version>
-            </script>
-            <script type="UnInstall">
-              <path>Providers\DataProviders\SqlDataProvider</path>
-              <name>Uninstall.SqlDataProvider</name>
-              <version>00.00.01</version>
-            </script>
-          </scripts>
-        </component>
-
-        <component type="ResourceFile">
-          <resourceFiles>
-            <basePath>DesktopModules/<%= extensionName %></basePath>
-            <resourceFile>
-              <name>Resources.zip</name>
-            </resourceFile>
-          </resourceFiles>
-        </component>
-
-        <component type="Module">
-          <desktopModule>
-            <moduleName><%= extensionName %></moduleName>
-            <foldername><%= extensionName %></foldername>
-            <businessControllerClass><%= fullNamespace %>.Components.FeatureController</businessControllerClass>
-            <supportedFeatures />
-            <moduleDefinitions>
-              <moduleDefinition>
-                <friendlyName><%= extensionName %></friendlyName>
-                <defaultCacheTime>0</defaultCacheTime>
-                <moduleControls>
-                  <moduleControl>
-                    <controlKey />
-                    <controlSrc>DesktopModules/<%= extensionName %>/View.html</controlSrc>
-                    <supportsPartialRendering>False</supportsPartialRendering>
-                    <controlTitle />
-                    <controlType>View</controlType>
-                    <iconFile />
-                    <helpUrl />
-                    <viewOrder>0</viewOrder>
-                  </moduleControl>
-                  <moduleControl>
-                    <controlKey>QuickSettings</controlKey>
-                    <controlSrc>DesktopModules/<%= extensionName %>/Settings.html</controlSrc>
-                    <supportsPartialRendering>False</supportsPartialRendering>
-                    <controlTitle><%= extensionName %> Settings</controlTitle>
-                    <controlType>Edit</controlType>
-                    <iconFile />
-                    <helpUrl />
-                    <viewOrder>0</viewOrder>
-                  </moduleControl>
-                </moduleControls>
-              </moduleDefinition>
-            </moduleDefinitions>
-          </desktopModule>
-        </component>
-        <component type="Assembly">
-          <assemblies>
-            <assembly>
-              <name><%= fullNamespace %>.dll</name>
-              <path>bin</path>
-            </assembly>
-            <assembly>
-              <name>EntityFramework.dll</name>
-              <path>bin</path>
-            </assembly>
-            <assembly>
-              <name>EntityFramework.SqlServer.dll</name>
-              <path>bin</path>
-            </assembly>
-          </assemblies>
-        </component>
-      </components>
-    </package>
-  </packages>
-</dotnetnuke>
