--- generators/spa-vue/templates/common/Properties/AssemblyInfo.cs
@@ -1,34 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("<%= extensionName %>")]
-[assembly: AssemblyDescription("<%= description %>")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("<%= namespace %>")]
-[assembly: AssemblyProduct("<%= extensionName %> Extension for DNN")]
-[assembly: AssemblyCopyright("Copyright © <%= currentYear %> <%= namespace %>")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("<%= guid %>")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Revision and Build Numbers 
-// by using the '*' as shown below:
-[assembly: AssemblyVersion("00.00.01.*")]
-[assembly: AssemblyFileVersion("00.00.01.*")]
