--- generators/spa-vue/templates/common/Providers/DataProviders/SqlDataProvider/00.00.01.SqlDataProvider
@@ -1,42 +1,0 @@
-﻿/************************************************************/
-/*****              SqlDataProvider                     *****/
-/*****                                                  *****/
-/*****                                                  *****/
-/***** Note: To manually execute this script you must   *****/
-/*****       perform a search and replace operation     *****/
-/*****       for {databaseOwner} and {objectQualifier}  *****/
-/*****                                                  *****/
-/************************************************************/
-
-IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'{databaseOwner}[{objectQualifier}<%= extensionName %>_Items]') AND type in (N'U'))
-DROP TABLE {databaseOwner}[{objectQualifier}<%= extensionName %>_Items]
-GO
-
-CREATE TABLE {databaseOwner}{objectQualifier}<%= extensionName %>_Items
-    (
-    ItemId int NOT NULL IDENTITY (1, 1),
-    ItemName nvarchar(MAX) NOT NULL,
-    ItemDescription nvarchar(MAX) NOT NULL,
-    AssignedUserId int NULL,
-    ModuleId int NOT NULL,
-    CreatedOnDate datetime NOT NULL,
-    CreatedByUserId int NOT NULL,
-    LastModifiedOnDate datetime NOT NULL,
-    LastModifiedByUserId int NOT NULL
-    )  ON [PRIMARY]
-     TEXTIMAGE_ON [PRIMARY]
-GO
-
-
-ALTER TABLE {databaseOwner}{objectQualifier}<%= extensionName %>_Items ADD CONSTRAINT
-    PK_{objectQualifier}<%= extensionName %>_Items PRIMARY KEY CLUSTERED 
-    (
-    ItemId
-    ) WITH( STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
-
-GO
-
-
-/************************************************************/
-/*****              SqlDataProvider                     *****/
-/************************************************************/
