--- generators/spa-vue/templates/common/Providers/DataProviders/SqlDataProvider/Uninstall.SqlDataProvider
@@ -1,3 +1,0 @@
-﻿IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'{databaseOwner}[{objectQualifier}<%= extensionName %>_Items]') AND type in (N'U'))
-DROP TABLE {databaseOwner}[{objectQualifier}<%= extensionName %>_Items]
-GO
