+++ generators/spa/templates/VueJS/common/App_LocalResources/Settings.resx
