+++ generators/spa/templates/VueJS/common/BuildScripts/MSBuild.Community.Tasks.Targets
