+++ generators/spa/templates/VueJS/common/BuildScripts/ModulePackage.targets
