+++ generators/spa/templates/VueJS/common/Components/DataProviderHelper.cs
