+++ generators/spa/templates/VueJS/common/Components/FeatureController.cs
