+++ generators/spa/templates/VueJS/common/Data/DnnDbConfiguration.cs
