+++ generators/spa/templates/VueJS/common/Data/Item.cs
