+++ generators/spa/templates/VueJS/common/Module.build
@@ -1,0 +1,51 @@
+<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <ModulePath><%= moduleName %></ModulePath>
+    <DNNFileName><%= moduleName %></DNNFileName>
+    <PackageName><%= fullNamespace %></PackageName>
+    <Extension>zip</Extension>
+    <BuildScriptsPath>$(MSBuildProjectDirectory)\BuildScripts</BuildScriptsPath>
+    <WebsitePath>$(MSBuildProjectDirectory)\..\..\Website</WebsitePath>
+    <WebsiteInstallPath>$(WebsitePath)\Install\Module</WebsiteInstallPath>
+    <FullModulePath>$(WebsitePath)\DesktopModules\$(ModulePath)</FullModulePath>
+  </PropertyGroup>
+  <Import Project="$(BuildScriptsPath)\ModulePackage.Targets" />
+  <Target Name="AfterBuild" DependsOnTargets="CopyBin;GetFiles;DebugProject;PackageModule">
+  </Target>
+  <Target Name="GetFiles">
+    <ItemGroup>
+      <Manifest Include="*.dnn;*.dnn5;*.dnn6" />
+      <TextFiles Include="*.txt" Exclude="releasenotes.txt;license.txt" />
+      <SourceFiles Include="*.css;*.htm;*.html" />
+      <Views Include="*.ascx;" />
+      <ResourceFiles Include="App_LocalResources/*.resx" />
+      <Images Include="Images/**/*.*" />
+      <JsFiles Include="Scripts/**/*.*" />
+      <Styles Include="Styles/**/*.*" />
+      <DataProvider Include="Providers/DataProviders/SqlDataProvider/*.SqlDataProvider" />
+      <Resources Include="@(Manifest);@(TextFiles);@(SourceFiles);@(Views);@(ResourceFiles);@(Images);@(JsFiles);@(Styles);@(DataProvider)" />
+    </ItemGroup>
+  </Target>
+  <ItemGroup>
+    <SymbolsInclude Include="License.txt" />
+    <SymbolsInclude Include="ReleaseNotes.txt" />
+  </ItemGroup>
+  <Target Name="CopyBin">
+    <ItemGroup>
+      <BinSourceInclude Include="$(MSBuildProjectDirectory)\bin\*.dll" Exclude="$(MSBuildProjectDirectory)\bin\DotNetNuke*;$(MSBuildProjectDirectory)\bin\Telerik*;$(MSBuildProjectDirectory)\bin\System.*;$(MSBuildProjectDirectory)\bin\Microsoft.*;$(MSBuildProjectDirectory)\bin\Newtonsoft.Json.*" />
+      <BinSourceInclude Include="$(MSBuildProjectDirectory)\bin\*.pdb" Exclude="$(MSBuildProjectDirectory)\bin\DotNetNuke*;$(MSBuildProjectDirectory)\bin\Telerik*;$(MSBuildProjectDirectory)\bin\System.*;$(MSBuildProjectDirectory)\bin\Microsoft.*;$(MSBuildProjectDirectory)\bin\Newtonsoft.Json.*" />
+    </ItemGroup>
+    <Copy SourceFiles="@(BinSourceInclude)" DestinationFolder="$(WebsitePath)\bin" />
+  </Target>
+  <Target Name="DebugProject" Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <Copy SourceFiles="@(Manifest)" DestinationFolder="$(FullModulePath)" />
+    <Copy SourceFiles="@(TextFiles)" DestinationFolder="$(FullModulePath)" />
+    <Copy SourceFiles="@(SourceFiles)" DestinationFolder="$(FullModulePath)" />
+    <Copy SourceFiles="@(Views)" DestinationFolder="$(FullModulePath)" />
+    <Copy SourceFiles="@(ResourceFiles)" DestinationFolder="$(FullModulePath)\App_LocalResources" />
+    <Copy SourceFiles="@(Images)" DestinationFolder="$(FullModulePath)\Images\%(RecursiveDir)" />
+    <Copy SourceFiles="@(JsFiles)" DestinationFolder="$(FullModulePath)\Scripts\%(RecursiveDir)" />
+    <Copy SourceFiles="@(Styles)" DestinationFolder="$(FullModulePath)\Styles\%(RecursiveDir)" />
+    <Copy SourceFiles="@(DataProvider)" DestinationFolder="$(FullModulePath)\Providers\DataProviders\SqlDataProvider" />
+  </Target>
+</Project>
