+++ generators/spa/templates/VueJS/common/Providers/DataProviders/SqlDataProvider/Uninstall.SqlDataProvider
@@ -1,0 +1,3 @@
+IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'{databaseOwner}[{objectQualifier}<%= moduleName %>_Items]') AND type in (N'U'))
+DROP TABLE {databaseOwner}[{objectQualifier}<%= moduleName %>_Items]
+GO
