+++ generators/spa/templates/VueJS/common/ReleaseNotes.txt
@@ -1,0 +1,10 @@
+<h3><%= moduleName %></h3>
+<p class="Contributor">
+	<%= company %> <br />
+	<a href="mailto:<%= emailAddy %>"><%= emailAddy %></a>
+</p>
+<hr/>
+<div class="License">
+	<h3>About the <%= moduleName %></h3>
+	<p>This module is intended as an example of how to build DNN Modules with VueJS.</p>
+</div>
