+++ generators/spa/templates/VueJS/common/Services/ViewModels/ItemViewModel.cs
