+++ generators/spa/templates/VueJS/common/Services/ViewModels/UserViewModel.cs
