+++ generators/spa/templates/VueJS/common/View.html
@@ -1,0 +1,68 @@
+[JavaScript:{ jsname: "JQuery" }]
+[JavaScript:{ path: "https://cdn.jsdelivr.net/npm/vue/dist/vue.js"}]
+[JavaScript:{ path: "~/Resources/Shared/scripts/dnn.jquery.js"}]
+[JavaScript:{ path: "~/DesktopModules/<%= moduleName %>/scripts/ItemView.js"}]
+
+[ModuleAction:{controlKey : "Edit", securityAccessLevel : "Edit",  titleKey: "EditModule", localResourceFile: "~/DesktopModules/<%= moduleName %>/App_LocalResources/View.resx" }]
+
+<div id="vue-app-wrapper-[ModuleContext:ModuleId]">
+    <div id="app-[ModuleContext:ModuleId]">
+        <div>
+            <a @click="loadItems" href="#">[Resx:{key:"Refresh"}]</a>
+            <a v-if="userCanAdd" @click="addItem" href="#">[Resx:{key:"EditModule"}]</a>
+        </div>
+        <ul class="tm_tl">
+            <li class="tm_t" v-for="(item, index) in items" :key="item.id">
+                <h3>{{item.name}}</h3>
+                <div class="tm_td">{{item.description}}</div>
+                <edit-component :moduleid="moduleid"
+                                :id="item.id"
+                                :name="item.name"
+                                :description="item.description"
+                                :canedit="item.canEdit"
+                                :assigned-user="item.assignedUser"
+                                :users="users"
+                                v-on:reload="loadItems"
+                                v-on:edit-cancelled="cancelAdd"></edit-component>
+            </li>
+        </ul>
+    </div>
+</div>
+
+<script type="x-template" id="edit-component-[ModuleContext:ModuleId]">
+    <div>
+        <div v-if="editMode || item.id <= 0" class="dnnForm dnnEditBasicSettings" id="dnnEditBasicSettings">
+            <fieldset>
+                <div class="dnnFormItem">
+                    <div><label for="itemName">[Resx:{key:"lblName"}]</label></div>
+                    <input id="itemName" type="text" v-model="item.name" />
+                </div>
+                <div class="dnnFormItem">
+                    <div><label for="itemDescription">[Resx:{key:"lblDescription"}]</label></div>
+                    <textarea id="itemDescription" cols="20" rows="5" v-model="item.description"></textarea>
+                </div>
+                <div class="dnnFormItem">
+                    <div><label for="itemUser">[Resx:{key:"lblAssignedUser"}]</label></div>
+                    <select v-model="item.assignedUser">
+                        <option v-for="user in users" v-bind:value="user.id">
+                            {{ user.name }}
+                        </option>
+                    </select>
+                </div>
+            </fieldset>
+            <a href="#" @click="saveItem" class="dnnPrimaryAction">[Resx:{key:"btnSubmit"}]</a>
+            <a href="#" @click="cancelEdit" class="dnnSecondaryAction">[Resx:{key:"btnCancel"}]</a>
+        </div>
+        <div v-if="item.id > 0 && item.canedit">
+            <a v-if="!editMode" @click="toggleEditMode" href="#">[Resx:{key:"EditItem"}]</a>
+            <a @click="deleteItem" href="#">[Resx:{key:"DeleteItem"}]</a>
+        </div>
+    </div>
+</script>
+
+<script language="javascript" type="text/javascript">
+    jQuery(function ($) {
+        <%= moduleName %>.InitApp([ModuleContext:ModuleId], '[ModuleContext:EditMode]' === 'True');
+    });
+
+</script>
