+++ generators/spa/templates/VueJS/common/packages.config
