+++ generators/spa/templates/VueJS/common/web.Debug.config
