+++ generators/spa/templates/VueJS/common/web.Release.config
