+++ generators/spa/templates/VueJS/common/web.config
